# compiler/natives/src/reflect: Optimize Swapper by swapping JS array elements directly (bypass reflection).