--- a/compiler/natives/src/reflect/swapper.go
+++ b/compiler/natives/src/reflect/swapper.go
@@ -2,13 +2,16 @@
 
 package reflect
 
+import "github.com/gopherjs/gopherjs/js"
+
 func Swapper(slice interface{}) func(i, j int) {
 	v := ValueOf(slice)
 	if v.Kind() != Slice {
 		panic(&ValueError{Method: "Swapper", Kind: v.Kind()})
 	}
 	// Fast path for slices of size 0 and 1. Nothing to swap.
-	switch v.Len() {
+	vLen := uint(v.Len())
+	switch vLen {
 	case 0:
 		return func(i, j int) { panic("reflect: slice index out of range") }
 	case 1:
@@ -18,12 +21,13 @@ func Swapper(slice interface{}) func(i, j int) {
 			}
 		}
 	}
-	tmp := New(v.Type().Elem()).Elem()
+	s := js.InternalObject(slice).Get("$array")
 	return func(i, j int) {
-		v1 := v.Index(i)
-		v2 := v.Index(j)
-		tmp.Set(v1)
-		v1.Set(v2)
-		v2.Set(tmp)
+		if uint(i) >= vLen || uint(j) >= vLen {
+			panic("reflect: slice index out of range")
+		}
+		tmp := s.Index(i)
+		s.SetIndex(i, s.Index(j))
+		s.SetIndex(j, tmp)
 	}
 }