… compiler/natives/fs_vfsdata.go