Sort package was very slow and reflect when transpiled is bloated.

Before natives were added
```
gopherjs test sort --run=Slice --bench=Slice
goos: linux
goarch: js
BenchmarkSortString1K_Slice 	     100	  14450000 ns/op
BenchmarkStableInt1K_Slice  	      20	  55000000 ns/op
BenchmarkSortInt64K_Slice   	       1	1486000000 ns/op
PASS
ok  	sort	4.913s
```

After natives were added
```
gopherjs test sort --run=Slice --bench=Slice -v
goos: linux
goarch: js
BenchmarkSortString1K_Slice 	    2000	    984000 ns/op
BenchmarkStableInt1K_Slice  	    2000	    993500 ns/op
BenchmarkSortInt64K_Slice   	      20	  84300000 ns/op
PASS
ok  	sort	6.442s
```

Result is around a one order of magnitude faster (and smaller generated code when using sort!)