@bign8 2018-04-02
I was thinking of adding the following shim to sort as well, mind if I add it to your PR?

```go
package sort

import "github.com/gopherjs/gopherjs/js"

// Strings sorts a slice of strings in increasing order.
func Strings(a []string) {
	js.InternalObject(a).Get("$array").Call("sort")
}
```