@bign8 2018-04-03
Nevermind, that breaks a bunch of non-sort tests.  I'll look into it and make another PR later.