@myitcv 2018-04-03
Looks 👍to me @lologarithm!

@bign8 - I don't think that will work because in Go world strings are UTF-8 and in Javascript world they are [UTF-16](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String). But string sorting will benefit generally from this change by the looks of things.