@lologarithm 2018-04-04
One thing about this bothers me though. The 'slice.go' file in the stdlib imports reflect. With these changes the reflect import is no longer needed. However, because the import is still in the AST during the compile, we transpile the reflect package as well. This adds ~150kb of extra code that isn't needed. 

I wonder if there is a way to determine if any dependencies are no longer needed after applying the natives. Maybe search the package AST for any remaining references to each import in the stdlib? Might increase compilation time a little but could result in reduced "binary size".

Any thoughts?