@shurcooL 2018-04-04
Hi @lologarithm,

Thanks a lot for sending this PR. This is an incredible performance improvement!

We'll definitely want to merge this. But there are some minor issues and opportunities to improve that I'd like to share.

There is one issue of correctness. The implementation, as is, breaks the `sort.Slice` contract. Specifically, the documentation of [`sort.Slice`](https://godoc.org/sort#Slice):

> The function panics if the provided interface is not a slice.

That currently doesn't happen. This program did not panic in my testing:

```Go
package main

import "sort"

func main() {
	sort.Slice("hello", nil)
	println("still here")
}
```

I think we can make this PR even better (and it resolves the point you've mentioned above about `reflect` no longer being needed). Here's how.

The absolute majority of the performance improvement comes from the improved implementation of `reflect.Swapper`, one that bypasses the extra `reflect` GopherJS machinery and swaps the underlying slice elements directly with JavaScript. The rest of the performance improvement (e.g., getting the slice length via `Get("$length").Int()`) is comparatively negligible.

We can factor out the `reflect.Swapper` improvement into `reflect` package itself:

```diff
diff --git a/compiler/natives/src/reflect/swapper.go b/compiler/natives/src/reflect/swapper.go
index b163fab..4d4af03 100644
--- a/compiler/natives/src/reflect/swapper.go
+++ b/compiler/natives/src/reflect/swapper.go
@@ -2,6 +2,8 @@
 
 package reflect
 
+import "github.com/gopherjs/gopherjs/js"
+
 func Swapper(slice interface{}) func(i, j int) {
 	v := ValueOf(slice)
 	if v.Kind() != Slice {
@@ -18,12 +20,10 @@ func Swapper(slice interface{}) func(i, j int) {
 			}
 		}
 	}
-	tmp := New(v.Type().Elem()).Elem()
+	s := js.InternalObject(slice).Get("$array")
 	return func(i, j int) {
-		v1 := v.Index(i)
-		v2 := v.Index(j)
-		tmp.Set(v1)
-		v1.Set(v2)
-		v2.Set(tmp)
+		tmp := s.Index(i)
+		s.SetIndex(i, s.Index(j))
+		s.SetIndex(j, tmp)
 	}
 }
```

That way, `sort` doesn't have to be modified at all, and it will correctly panic on non-slice input. It also benefits everything that uses `reflect.Swapper`, not just those 3 functions in `sort`. I've benchmarked the result, and it's equally fast.

Here's the benchmark comparison of current vs this PR as is:

```
benchmark                       old ns/op      new ns/op     delta
BenchmarkSortString1K_Slice     17310000       510000        -97.05%
BenchmarkStableInt1K_Slice      84900000       793000        -99.07%
BenchmarkSortInt64K_Slice       2241000000     37200000      -98.34%
```

Here's the benchmark comparison of current vs my proposed change:

```
benchmark                       old ns/op      new ns/op     delta
BenchmarkSortString1K_Slice     17310000       531000        -96.93%
BenchmarkStableInt1K_Slice      84900000       788000        -99.07%
BenchmarkSortInt64K_Slice       2241000000     36566666      -98.37%
```

@lologarithm What do you think?