@lologarithm 2018-04-04
I like it! Should I open a new PR or just pull those changes in here and remove the changes to slice.go?