@shurcooL 2018-04-04
It'd be fine to modify this existing PR. Just push as new commits (it'll all be squashed when merging) so the history is preserved.

I've realized that we'll likely need to add a bound check to the swapper func, similar to one in [Go stdlib](https://gotools.org/reflect?rev=go1.10.1#swapper.go-L65-L67):

```Go
if uint(i) >= uint(s.Len) || uint(j) >= uint(s.Len) {
	panic("reflect: slice index out of range")
}
```

Lemme see how much it affects performance...