@shurcooL 2018-04-04
You have to be careful not to call `v.Len()` inside the swapper func, because that's expensive:

```
$ benchcmp before.txt after.txt 
benchmark                       old ns/op     new ns/op     delta
BenchmarkSortString1K_Slice     528000        3826666       +624.75%
BenchmarkStableInt1K_Slice      823000        12430000      +1410.33%
BenchmarkSortInt64K_Slice       37300000      335000000     +798.12%
```

But after factoring out that call (the passed slice size doesn't change), it's not expensive at all:

```
$ benchcmp before.txt after2.txt 
benchmark                       old ns/op     new ns/op     delta
BenchmarkSortString1K_Slice     528000        535500        +1.42%
BenchmarkStableInt1K_Slice      823000        812500        -1.28%
BenchmarkSortInt64K_Slice       37300000      37400000      +0.27%
```

This was the final version I tested:

```Go
// +build js

package reflect

import "github.com/gopherjs/gopherjs/js"

func Swapper(slice interface{}) func(i, j int) {
	v := ValueOf(slice)
	if v.Kind() != Slice {
		panic(&ValueError{Method: "Swapper", Kind: v.Kind()})
	}
	// Fast path for slices of size 0 and 1. Nothing to swap.
	vLen := uint(v.Len())
	switch vLen {
	case 0:
		return func(i, j int) { panic("reflect: slice index out of range") }
	case 1:
		return func(i, j int) {
			if i != 0 || j != 0 {
				panic("reflect: slice index out of range")
			}
		}
	}
	s := js.InternalObject(slice).Get("$array")
	return func(i, j int) {
		if uint(i) >= vLen || uint(j) >= vLen {
			panic("reflect: slice index out of range")
		}
		tmp := s.Index(i)
		s.SetIndex(i, s.Index(j))
		s.SetIndex(j, tmp)
	}
}
```