@lologarithm 2018-04-04
I benched around 1-5% slowdown which makes this still quite a net positive