@lologarithm 2018-04-04
I thin yours is better -- one less call to size. ill update