@lologarithm 2018-04-04
rebased to squash all the commits down