@shurcooL 2018-04-04
> rebased to squash all the commits down

I wish you hadn't, it'd be nice to be able to see the history of the PR (unfortunately, unlike Gerrit, GitHub PRs don't track anything past force pushes). But it's okay.